// frontend/src/core/api/generated/model/setupInviteView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Constructor ERP API
 * OpenAPI spec version: 1.0
 */

export interface SetupInviteView {
  companyName: string
  /** @nullable */
  phone?: string | null
  /** @nullable */
  email?: string | null
  /** @nullable */
  city?: string | null
  /** @nullable */
  logoUrl?: string | null
  /** @nullable */
  planName?: string | null
  /** @nullable */
  endsOn?: string | null
  features: string[]
}